requirements.txt
cryptography
curl_cffi
mmh3
pycryptodome